requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coverage==7.2.7
Django==4.2.3
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==23.3
django-rest-framework-nested==0.0.1
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
excel-base==1.0.4
exceptiongroup==1.1.2
flake8==6.0.0
flake8-html==0.4.3
idna==3.4
iniconfig==2.0.0
isoweek==1.3.3
Jinja2==3.1.2
language==0.6
Markdown==3.5
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==1.7.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
xlwt==1.3.0
